+++ CHANGELOG.md
@@ -12,6 +12,7 @@ If you are upgrading from a version prior to 8.12.10, following the upgrade, you
 
 ### Features
 
+ - Add camera library: libcamera
  - Add ability to duplicate Inputs
  - Add ability to invert the saved duty cycle for PWM Outputs